+++ client/git-timestamp.py
@@ -1,3 +1,0 @@
-#!/usr/bin/python
-from igitt_client import timestamp
-timestamp.main()