+++ git-timestamp.py
@@ -1,0 +1,3 @@
+#!/usr/bin/python
+from git_timestamp import timestamp
+timestamp.main()