+++ git_timestamp/timestamp.py
@@ -35,7 +35,7 @@ import gnupg
 import pygit2 as git
 import requests
 
-VERSION = '0.9.2+'
+VERSION = '0.9.3'
 
 
 class GitArgumentParser(argparse.ArgumentParser):