+++ CHANGELOG.md
@@ -6,8 +6,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 # 0.9.2+ - [Unreleased]
 ## Added
+### Client
+- Default server set to `https://gitta.enotar.ch`; can be changed with
+  `git config [--global] timestamper.server …`
+- Allow dots in tag/branch names, as long as they are not next to each other
+  (i.e., `..` is not allowed)
+
 ## Fixed
+
 ## Changed
+- Split into client (git-timestamp) and server (zeitgitterd).
+
+### Client
+- Persistent information about the timestampers' keys is now stored
+  in the global git configuration (key also stored globally; more TOFU-like)
+
 
 # 0.9.2 - 2019-05-10
 ## Added