+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-# 0.9.2+ - [Unreleased]
+# 0.9.3 - 2019-05-12
 ## Added
 - Default server set to `https://gitta.enotar.ch`; can be changed with
   `git config [--global] timestamper.server …`